requirements.txt
azure-ai-vision-imageanalysis
azure-core
pillow
python-dotenv
requests
